GAL-18 review fix
