+++ packages/client/src/pages/Forum/index.module.scss
@@ -5,7 +5,7 @@
   font-size: var(--font-size-20);
 
   @include page-container;
-  @include leederBoard-background;
+  @include leaderBoard-background;
   @include grid-template;
 }
