+++ packages/client/src/pages/LeaderBoard/index.module.scss
@@ -1,7 +1,7 @@
 @import '../../style/backgrounds';
 
 .leaderboardPage {
-  @include leederBoard-background;
+  @include leaderBoard-background;
   padding: 85px 104px;
   min-height: 100vh;
   min-width: 100vw;
