+++ packages/client/src/pages/Registration/index.module.scss
@@ -1,5 +1,5 @@
 @import '@/style/backgrounds';
 
 .registration {
-  @include login-background;
+  @include system-background;
 }
